[refactor] changes project structure
- Back to springboot after micronaut not working.
- Added a folder for the crawler
- Changes api to search-api to encapsulate funcionalities
