+++ crawler/.folder
@@ -1,0 +1,1 @@
+Initialization of the folder for the crawler
