+++ docker-compose.yml
@@ -2,6 +2,6 @@ version: '3'
 
 services:
   api:
-    build: ./api
+    build: ./search-api
     ports:
       - "8082:8082"
