+++ search-api/src/main/java/org/weso/snoicd/persistence/Persistence.java
@@ -7,7 +7,7 @@
  * See /LICENSE for license information.
  * 
  */
-package org.weso.snoicd.repositories;
+package org.weso.snoicd.persistence;
 
 import java.util.Set;
