+++ search-api/src/main/java/org/weso/snoicd/persistence/WarmUpMemory.java
@@ -7,7 +7,7 @@
  * See /LICENSE for license information.
  * 
  */
-package org.weso.snoicd.repositories;
+package org.weso.snoicd.persistence;
 
 import java.io.File;
 import java.io.IOException;
