+++ search-api/src/main/java/org/weso/snoicd/search/AllFieldsSearch.java
