+++ search-api/src/main/java/org/weso/snoicd/search/InvalidSearch.java
