+++ search-api/src/main/java/org/weso/snoicd/swagger/SwaggerConfig.java
