+++ search-api/src/main/java/org/weso/snoicd/types/ResponseToQuery.java
