+++ search-api/src/main/java/org/weso/snoicd/utils/StringNormalizatior.java
