+++ search-api/src/main/scripts/org/weso/snoicd/scripts/package.info
