+++ search-api/src/test/java/org/weso/snoicd/archTest/CodingRulesTest.java
