+++ search-api/src/test/java/org/weso/snoicd/archTest/NamingConventionTest.java
