+++ search-api/src/test/java/org/weso/snoicd/types/ResponseToQuerySmokeTest.java
